+++ great_expectations/data_context/data_context/abstract_data_context.py
@@ -1,4 +1,51 @@
-from abc import ABC
+import configparser
+import copy
+import json
+import logging
+import os
+from abc import ABC, abstractmethod
+from typing import Dict, List, Mapping, Optional, Union
+
+import great_expectations.exceptions.exceptions as ge_exceptions
+from great_expectations.core.config_peer import ConfigPeer
+from great_expectations.core.yaml_handler import YAMLHandler
+from great_expectations.data_context.types.base import (
+    AnonymizedUsageStatisticsConfig,
+    DataContextConfig,
+    DataContextConfigDefaults,
+    anonymizedUsageStatisticsSchema,
+)
+from great_expectations.data_context.util import (
+    PasswordMasker,
+    build_store_from_config,
+    instantiate_class_from_config,
+    substitute_all_config_variables,
+)
+from great_expectations.datasource import BaseDatasource, LegacyDatasource
+from great_expectations.rule_based_profiler.data_assistant.data_assistant_dispatcher import (
+    DataAssistantDispatcher,
+)
+
+from great_expectations.data_context.store import (  # isort:skip
+    ExpectationsStore,
+    ValidationsStore,
+)
+
+from great_expectations.core.usage_statistics.usage_statistics import (  # isort: skip
+    UsageStatisticsHandler,
+)
+
+from great_expectations.data_context.store import Store  # isort:skip
+from great_expectations.data_context.store import TupleStoreBackend  # isort:skip
+
+# this is here because of circular imports
+from great_expectations.rule_based_profiler import RuleBasedProfiler  # isort:skip
+from great_expectations.rule_based_profiler import RuleBasedProfilerResult  # isort:skip
+
+
+logger = logging.getLogger(__name__)
+
+yaml: YAMLHandler = YAMLHandler()
 
 
 class AbstractDataContext(ABC):
@@ -11,4 +58,210 @@ class AbstractDataContext(ABC):
     TODO: eventually the dependency on ConfigPeer will be removed and this will become a pure ABC.
     """
 
-    pass
+    PROFILING_ERROR_CODE_TOO_MANY_DATA_ASSETS = 2
+    PROFILING_ERROR_CODE_SPECIFIED_DATA_ASSETS_NOT_FOUND = 3
+    PROFILING_ERROR_CODE_NO_BATCH_KWARGS_GENERATORS_FOUND = 4
+    PROFILING_ERROR_CODE_MULTIPLE_BATCH_KWARGS_GENERATORS_FOUND = 5
+
+    DOLLAR_SIGN_ESCAPE_STRING = r"\$"
+    FALSEY_STRINGS = ["FALSE", "false", "False", "f", "F", "0"]
+
+    TEST_YAML_CONFIG_SUPPORTED_STORE_TYPES = [
+        "ExpectationsStore",
+        "ValidationsStore",
+        "HtmlSiteStore",
+        "EvaluationParameterStore",
+        "MetricStore",
+        "SqlAlchemyQueryStore",
+        "CheckpointStore",
+        "ProfilerStore",
+    ]
+    TEST_YAML_CONFIG_SUPPORTED_DATASOURCE_TYPES = [
+        "Datasource",
+        "SimpleSqlalchemyDatasource",
+    ]
+    TEST_YAML_CONFIG_SUPPORTED_DATA_CONNECTOR_TYPES = [
+        "InferredAssetFilesystemDataConnector",
+        "ConfiguredAssetFilesystemDataConnector",
+        "InferredAssetS3DataConnector",
+        "ConfiguredAssetS3DataConnector",
+        "InferredAssetAzureDataConnector",
+        "ConfiguredAssetAzureDataConnector",
+        "InferredAssetGCSDataConnector",
+        "ConfiguredAssetGCSDataConnector",
+        "InferredAssetSqlDataConnector",
+        "ConfiguredAssetSqlDataConnector",
+    ]
+    TEST_YAML_CONFIG_SUPPORTED_CHECKPOINT_TYPES = [
+        "Checkpoint",
+        "SimpleCheckpoint",
+    ]
+    TEST_YAML_CONFIG_SUPPORTED_PROFILER_TYPES = [
+        "RuleBasedProfiler",
+    ]
+    ALL_TEST_YAML_CONFIG_DIAGNOSTIC_INFO_TYPES = [
+        "__substitution_error__",
+        "__yaml_parse_error__",
+        "__custom_subclass_not_core_ge__",
+        "__class_name_not_provided__",
+    ]
+    ALL_TEST_YAML_CONFIG_SUPPORTED_TYPES = (
+        TEST_YAML_CONFIG_SUPPORTED_STORE_TYPES
+        + TEST_YAML_CONFIG_SUPPORTED_DATASOURCE_TYPES
+        + TEST_YAML_CONFIG_SUPPORTED_DATA_CONNECTOR_TYPES
+        + TEST_YAML_CONFIG_SUPPORTED_CHECKPOINT_TYPES
+        + TEST_YAML_CONFIG_SUPPORTED_PROFILER_TYPES
+    )
+
+    UNCOMMITTED_DIRECTORIES = ["data_docs", "validations"]
+    GE_UNCOMMITTED_DIR = "uncommitted"
+    BASE_DIRECTORIES = [
+        DataContextConfigDefaults.CHECKPOINTS_BASE_DIRECTORY.value,
+        DataContextConfigDefaults.EXPECTATIONS_BASE_DIRECTORY.value,
+        DataContextConfigDefaults.PLUGINS_BASE_DIRECTORY.value,
+        DataContextConfigDefaults.PROFILERS_BASE_DIRECTORY.value,
+        GE_UNCOMMITTED_DIR,
+    ]
+    GE_DIR = "great_expectations"
+    GE_YML = "great_expectations.yml"
+    GE_EDIT_NOTEBOOK_DIR = GE_UNCOMMITTED_DIR
+
+    GLOBAL_CONFIG_PATHS = [
+        os.path.expanduser("~/.great_expectations/great_expectations.conf"),
+        "/etc/great_expectations.conf",
+    ]
+
+    _data_context = None
+
+    def __init__(
+        self,
+        project_config: Union[DataContextConfig, Mapping],
+        runtime_environment: Optional[dict] = None
+        # context_root_dir do we keep this?
+    ):
+        self._project_config = project_config
+        # is this the property that we want?
+        self._project_config_with_variables_substituted: Optional[
+            DataContextConfig
+        ] = None
+
+        self.runtime_environment = runtime_environment or {}
+
+        #
+        self._apply_global_config_overrides()
+
+    ### properties
+    @property
+    def config(self) -> DataContextConfig:
+        return self._project_config
+
+    ### private methods
+    def _apply_global_config_overrides(self) -> None:
+        # check for global usage statistics opt out
+        validation_errors = {}
+
+        if self._check_global_usage_statistics_opt_out():
+            logger.info(
+                "Usage statistics is disabled globally. Applying override to project_config."
+            )
+            self.config.anonymous_usage_statistics.enabled = False
+
+        # check for global data_context_id
+        global_data_context_id = self._get_global_config_value(
+            environment_variable="GE_DATA_CONTEXT_ID",
+            conf_file_section="anonymous_usage_statistics",
+            conf_file_option="data_context_id",
+        )
+        if global_data_context_id:
+            data_context_id_errors = anonymizedUsageStatisticsSchema.validate(
+                {"data_context_id": global_data_context_id}
+            )
+            if not data_context_id_errors:
+                logger.info(
+                    "data_context_id is defined globally. Applying override to project_config."
+                )
+                self.config.anonymous_usage_statistics.data_context_id = (
+                    global_data_context_id
+                )
+            else:
+                validation_errors.update(data_context_id_errors)
+        # check for global usage_statistics url
+        global_usage_statistics_url = self._get_global_config_value(
+            environment_variable="GE_USAGE_STATISTICS_URL",
+            conf_file_section="anonymous_usage_statistics",
+            conf_file_option="usage_statistics_url",
+        )
+        if global_usage_statistics_url:
+            usage_statistics_url_errors = anonymizedUsageStatisticsSchema.validate(
+                {"usage_statistics_url": global_usage_statistics_url}
+            )
+            if not usage_statistics_url_errors:
+                logger.info(
+                    "usage_statistics_url is defined globally. Applying override to project_config."
+                )
+                self.config.anonymous_usage_statistics.usage_statistics_url = (
+                    global_usage_statistics_url
+                )
+            else:
+                validation_errors.update(usage_statistics_url_errors)
+        if validation_errors:
+            logger.warning(
+                "The following globally-defined config variables failed validation:\n{}\n\n"
+                "Please fix the variables if you would like to apply global values to project_config.".format(
+                    json.dumps(validation_errors, indent=2)
+                )
+            )
+
+    # static methods
+
+    @staticmethod
+    def _check_global_usage_statistics_opt_out() -> bool:
+        if os.environ.get("GE_USAGE_STATS", False):
+            ge_usage_stats = os.environ.get("GE_USAGE_STATS")
+            if ge_usage_stats in AbstractDataContext.FALSEY_STRINGS:
+                return True
+            else:
+                logger.warning(
+                    "GE_USAGE_STATS environment variable must be one of: {}".format(
+                        AbstractDataContext.FALSEY_STRINGS
+                    )
+                )
+        for config_path in AbstractDataContext.GLOBAL_CONFIG_PATHS:
+            config = configparser.ConfigParser()
+            states = config.BOOLEAN_STATES
+            for falsey_string in AbstractDataContext.FALSEY_STRINGS:
+                states[falsey_string] = False
+            states["TRUE"] = True
+            states["True"] = True
+            config.BOOLEAN_STATES = states
+            config.read(config_path)
+            try:
+                if config.getboolean("anonymous_usage_statistics", "enabled") is False:
+                    # If stats are disabled, then opt out is true
+                    return True
+            except (ValueError, configparser.Error):
+                pass
+
+    #### class method
+    @classmethod
+    def _get_global_config_value(
+        cls,
+        environment_variable: Optional[str] = None,
+        conf_file_section=None,
+        conf_file_option=None,
+    ) -> Optional[str]:
+        assert (conf_file_section and conf_file_option) or (
+            not conf_file_section and not conf_file_option
+        ), "Must pass both 'conf_file_section' and 'conf_file_option' or neither."
+        if environment_variable and os.environ.get(environment_variable, False):
+            return os.environ.get(environment_variable)
+        if conf_file_section and conf_file_option:
+            for config_path in AbstractDataContext.GLOBAL_CONFIG_PATHS:
+                config = configparser.ConfigParser()
+                config.read(config_path)
+                config_value = config.get(
+                    conf_file_section, conf_file_option, fallback=None
+                )
+                if config_value:
+                    return config_value
+        return None
