+++ great_expectations/data_context/data_context/base_data_context.py
@@ -239,76 +239,7 @@ class BaseDataContext(EphemeralDataContext, ConfigPeer):
     --ge-feature-maturity-info--
     """
 
-    PROFILING_ERROR_CODE_TOO_MANY_DATA_ASSETS = 2
-    PROFILING_ERROR_CODE_SPECIFIED_DATA_ASSETS_NOT_FOUND = 3
-    PROFILING_ERROR_CODE_NO_BATCH_KWARGS_GENERATORS_FOUND = 4
-    PROFILING_ERROR_CODE_MULTIPLE_BATCH_KWARGS_GENERATORS_FOUND = 5
-    UNCOMMITTED_DIRECTORIES = ["data_docs", "validations"]
-    GE_UNCOMMITTED_DIR = "uncommitted"
-    BASE_DIRECTORIES = [
-        DataContextConfigDefaults.CHECKPOINTS_BASE_DIRECTORY.value,
-        DataContextConfigDefaults.EXPECTATIONS_BASE_DIRECTORY.value,
-        DataContextConfigDefaults.PLUGINS_BASE_DIRECTORY.value,
-        DataContextConfigDefaults.PROFILERS_BASE_DIRECTORY.value,
-        GE_UNCOMMITTED_DIR,
-    ]
-    GE_DIR = "great_expectations"
-    GE_YML = "great_expectations.yml"
-    GE_EDIT_NOTEBOOK_DIR = GE_UNCOMMITTED_DIR
-    FALSEY_STRINGS = ["FALSE", "false", "False", "f", "F", "0"]
-    GLOBAL_CONFIG_PATHS = [
-        os.path.expanduser("~/.great_expectations/great_expectations.conf"),
-        "/etc/great_expectations.conf",
-    ]
-    DOLLAR_SIGN_ESCAPE_STRING = r"\$"
-    TEST_YAML_CONFIG_SUPPORTED_STORE_TYPES = [
-        "ExpectationsStore",
-        "ValidationsStore",
-        "HtmlSiteStore",
-        "EvaluationParameterStore",
-        "MetricStore",
-        "SqlAlchemyQueryStore",
-        "CheckpointStore",
-        "ProfilerStore",
-    ]
-    TEST_YAML_CONFIG_SUPPORTED_DATASOURCE_TYPES = [
-        "Datasource",
-        "SimpleSqlalchemyDatasource",
-    ]
-    TEST_YAML_CONFIG_SUPPORTED_DATA_CONNECTOR_TYPES = [
-        "InferredAssetFilesystemDataConnector",
-        "ConfiguredAssetFilesystemDataConnector",
-        "InferredAssetS3DataConnector",
-        "ConfiguredAssetS3DataConnector",
-        "InferredAssetAzureDataConnector",
-        "ConfiguredAssetAzureDataConnector",
-        "InferredAssetGCSDataConnector",
-        "ConfiguredAssetGCSDataConnector",
-        "InferredAssetSqlDataConnector",
-        "ConfiguredAssetSqlDataConnector",
-    ]
-    TEST_YAML_CONFIG_SUPPORTED_CHECKPOINT_TYPES = [
-        "Checkpoint",
-        "SimpleCheckpoint",
-    ]
-    TEST_YAML_CONFIG_SUPPORTED_PROFILER_TYPES = [
-        "RuleBasedProfiler",
-    ]
-    ALL_TEST_YAML_CONFIG_DIAGNOSTIC_INFO_TYPES = [
-        "__substitution_error__",
-        "__yaml_parse_error__",
-        "__custom_subclass_not_core_ge__",
-        "__class_name_not_provided__",
-    ]
-    ALL_TEST_YAML_CONFIG_SUPPORTED_TYPES = (
-        TEST_YAML_CONFIG_SUPPORTED_STORE_TYPES
-        + TEST_YAML_CONFIG_SUPPORTED_DATASOURCE_TYPES
-        + TEST_YAML_CONFIG_SUPPORTED_DATA_CONNECTOR_TYPES
-        + TEST_YAML_CONFIG_SUPPORTED_CHECKPOINT_TYPES
-        + TEST_YAML_CONFIG_SUPPORTED_PROFILER_TYPES
-    )
-
-    _data_context = None
+    DOLLAR_SIGN_ESCAPE_STRING = r"/$"
 
     @classmethod
     def validate_config(cls, project_config: Union[DataContextConfig, Mapping]) -> bool:
@@ -342,6 +273,10 @@ class BaseDataContext(EphemeralDataContext, ConfigPeer):
         Returns:
             None
         """
+        super().__init__(
+            project_config=project_config, runtime_environment=runtime_environment
+        )
+
         if not BaseDataContext.validate_config(project_config):
             raise ge_exceptions.InvalidConfigError(
                 "Your project_config is not valid. Try using the CLI check-config command."
@@ -481,113 +416,113 @@ class BaseDataContext(EphemeralDataContext, ConfigPeer):
                 # caught at the context.get_batch() step. So we just pass here.
                 pass
 
-    def _apply_global_config_overrides(self) -> None:
-        # check for global usage statistics opt out
-        validation_errors = {}
+    # def _apply_global_config_overrides(self) -> None:
+    #     # check for global usage statistics opt out
+    #     validation_errors = {}
+    #
+    #     if self._check_global_usage_statistics_opt_out():
+    #         logger.info(
+    #             "Usage statistics is disabled globally. Applying override to project_config."
+    #         )
+    #         self.config.anonymous_usage_statistics.enabled = False
+    #
+    #     # check for global data_context_id
+    #     global_data_context_id = self._get_global_config_value(
+    #         environment_variable="GE_DATA_CONTEXT_ID",
+    #         conf_file_section="anonymous_usage_statistics",
+    #         conf_file_option="data_context_id",
+    #     )
+    #     if global_data_context_id:
+    #         data_context_id_errors = anonymizedUsageStatisticsSchema.validate(
+    #             {"data_context_id": global_data_context_id}
+    #         )
+    #         if not data_context_id_errors:
+    #             logger.info(
+    #                 "data_context_id is defined globally. Applying override to project_config."
+    #             )
+    #             self.config.anonymous_usage_statistics.data_context_id = (
+    #                 global_data_context_id
+    #             )
+    #         else:
+    #             validation_errors.update(data_context_id_errors)
+    #     # check for global usage_statistics url
+    #     global_usage_statistics_url = self._get_global_config_value(
+    #         environment_variable="GE_USAGE_STATISTICS_URL",
+    #         conf_file_section="anonymous_usage_statistics",
+    #         conf_file_option="usage_statistics_url",
+    #     )
+    #     if global_usage_statistics_url:
+    #         usage_statistics_url_errors = anonymizedUsageStatisticsSchema.validate(
+    #             {"usage_statistics_url": global_usage_statistics_url}
+    #         )
+    #         if not usage_statistics_url_errors:
+    #             logger.info(
+    #                 "usage_statistics_url is defined globally. Applying override to project_config."
+    #             )
+    #             self.config.anonymous_usage_statistics.usage_statistics_url = (
+    #                 global_usage_statistics_url
+    #             )
+    #         else:
+    #             validation_errors.update(usage_statistics_url_errors)
+    #     if validation_errors:
+    #         logger.warning(
+    #             "The following globally-defined config variables failed validation:\n{}\n\n"
+    #             "Please fix the variables if you would like to apply global values to project_config.".format(
+    #                 json.dumps(validation_errors, indent=2)
+    #             )
+    #         )
 
-        if self._check_global_usage_statistics_opt_out():
-            logger.info(
-                "Usage statistics is disabled globally. Applying override to project_config."
-            )
-            self.config.anonymous_usage_statistics.enabled = False
+    # @classmethod
+    # def _get_global_config_value(
+    #     cls,
+    #     environment_variable: Optional[str] = None,
+    #     conf_file_section=None,
+    #     conf_file_option=None,
+    # ) -> Optional[str]:
+    #     assert (conf_file_section and conf_file_option) or (
+    #         not conf_file_section and not conf_file_option
+    #     ), "Must pass both 'conf_file_section' and 'conf_file_option' or neither."
+    #     if environment_variable and os.environ.get(environment_variable, False):
+    #         return os.environ.get(environment_variable)
+    #     if conf_file_section and conf_file_option:
+    #         for config_path in BaseDataContext.GLOBAL_CONFIG_PATHS:
+    #             config = configparser.ConfigParser()
+    #             config.read(config_path)
+    #             config_value = config.get(
+    #                 conf_file_section, conf_file_option, fallback=None
+    #             )
+    #             if config_value:
+    #                 return config_value
+    #     return None
 
-        # check for global data_context_id
-        global_data_context_id = self._get_global_config_value(
-            environment_variable="GE_DATA_CONTEXT_ID",
-            conf_file_section="anonymous_usage_statistics",
-            conf_file_option="data_context_id",
-        )
-        if global_data_context_id:
-            data_context_id_errors = anonymizedUsageStatisticsSchema.validate(
-                {"data_context_id": global_data_context_id}
-            )
-            if not data_context_id_errors:
-                logger.info(
-                    "data_context_id is defined globally. Applying override to project_config."
-                )
-                self.config.anonymous_usage_statistics.data_context_id = (
-                    global_data_context_id
-                )
-            else:
-                validation_errors.update(data_context_id_errors)
-        # check for global usage_statistics url
-        global_usage_statistics_url = self._get_global_config_value(
-            environment_variable="GE_USAGE_STATISTICS_URL",
-            conf_file_section="anonymous_usage_statistics",
-            conf_file_option="usage_statistics_url",
-        )
-        if global_usage_statistics_url:
-            usage_statistics_url_errors = anonymizedUsageStatisticsSchema.validate(
-                {"usage_statistics_url": global_usage_statistics_url}
-            )
-            if not usage_statistics_url_errors:
-                logger.info(
-                    "usage_statistics_url is defined globally. Applying override to project_config."
-                )
-                self.config.anonymous_usage_statistics.usage_statistics_url = (
-                    global_usage_statistics_url
-                )
-            else:
-                validation_errors.update(usage_statistics_url_errors)
-        if validation_errors:
-            logger.warning(
-                "The following globally-defined config variables failed validation:\n{}\n\n"
-                "Please fix the variables if you would like to apply global values to project_config.".format(
-                    json.dumps(validation_errors, indent=2)
-                )
-            )
-
-    @classmethod
-    def _get_global_config_value(
-        cls,
-        environment_variable: Optional[str] = None,
-        conf_file_section=None,
-        conf_file_option=None,
-    ) -> Optional[str]:
-        assert (conf_file_section and conf_file_option) or (
-            not conf_file_section and not conf_file_option
-        ), "Must pass both 'conf_file_section' and 'conf_file_option' or neither."
-        if environment_variable and os.environ.get(environment_variable, False):
-            return os.environ.get(environment_variable)
-        if conf_file_section and conf_file_option:
-            for config_path in BaseDataContext.GLOBAL_CONFIG_PATHS:
-                config = configparser.ConfigParser()
-                config.read(config_path)
-                config_value = config.get(
-                    conf_file_section, conf_file_option, fallback=None
-                )
-                if config_value:
-                    return config_value
-        return None
-
-    @staticmethod
-    def _check_global_usage_statistics_opt_out() -> bool:
-        if os.environ.get("GE_USAGE_STATS", False):
-            ge_usage_stats = os.environ.get("GE_USAGE_STATS")
-            if ge_usage_stats in BaseDataContext.FALSEY_STRINGS:
-                return True
-            else:
-                logger.warning(
-                    "GE_USAGE_STATS environment variable must be one of: {}".format(
-                        BaseDataContext.FALSEY_STRINGS
-                    )
-                )
-        for config_path in BaseDataContext.GLOBAL_CONFIG_PATHS:
-            config = configparser.ConfigParser()
-            states = config.BOOLEAN_STATES
-            for falsey_string in BaseDataContext.FALSEY_STRINGS:
-                states[falsey_string] = False
-            states["TRUE"] = True
-            states["True"] = True
-            config.BOOLEAN_STATES = states
-            config.read(config_path)
-            try:
-                if config.getboolean("anonymous_usage_statistics", "enabled") is False:
-                    # If stats are disabled, then opt out is true
-                    return True
-            except (ValueError, configparser.Error):
-                pass
-        return False
+    # @staticmethod
+    # def _check_global_usage_statistics_opt_out() -> bool:
+    #     if os.environ.get("GE_USAGE_STATS", False):
+    #         ge_usage_stats = os.environ.get("GE_USAGE_STATS")
+    #         if ge_usage_stats in BaseDataContext.FALSEY_STRINGS:
+    #             return True
+    #         else:
+    #             logger.warning(
+    #                 "GE_USAGE_STATS environment variable must be one of: {}".format(
+    #                     BaseDataContext.FALSEY_STRINGS
+    #                 )
+    #             )
+    #     for config_path in BaseDataContext.GLOBAL_CONFIG_PATHS:
+    #         config = configparser.ConfigParser()
+    #         states = config.BOOLEAN_STATES
+    #         for falsey_string in BaseDataContext.FALSEY_STRINGS:
+    #             states[falsey_string] = False
+    #         states["TRUE"] = True
+    #         states["True"] = True
+    #         config.BOOLEAN_STATES = states
+    #         config.read(config_path)
+    #         try:
+    #             if config.getboolean("anonymous_usage_statistics", "enabled") is False:
+    #                 # If stats are disabled, then opt out is true
+    #                 return True
+    #         except (ValueError, configparser.Error):
+    #             pass
+    #     return False
 
     def _construct_data_context_id(self) -> str:
         """
